hackaton

+++ server/api/product/changeProductPreferences.post.ts
@@ -3,32 +3,44 @@ import prisma from "~/lib/prisma";
 export default defineEventHandler(async (event) => {
     const body = await readBody(event)
 
-    const organization_ = await prisma.organization.findUnique({
+    const organization_ = await prisma.organization.findFirst({
         where: {
-            userId: event.context.user.id,
+            User: {
+                id: event.context.user.id,
+            }
         },
     })
 
-    await prisma.interestedProduct.deleteMany({
-        where: {
-            organizationId: organization_!.id,
-        },
-    })
+    if (organization_ == null) {
+        setResponseStatus(event, 403)
+        return
+    }
 
-    await prisma.watchedProduct.deleteMany({
-        where: {
-            organizationId: organization_!.id,
-        },
-    })
+    if(body.interestedProductIds != undefined){
+        await prisma.interestedProduct.deleteMany({
+            where: {
+                organizationId: organization_!.id,
+            },
+        })
+        await prisma.interestedProduct.createMany({
+            data: body.interestedProductIds.map((e: number)=>{return  {organizationId: organization_?.id, productId: e}})
+        })
+    }
 
-    await prisma.interestedProduct.createMany({
-        data: body.interestedProductIds.map((e: number)=>{return  {organizationId: organization_?.id, productId: e}})
-    })
+    if (body.watchedProductIds != undefined) {
+        await prisma.watchedProduct.deleteMany({
+            where: {
+                organizationId: organization_!.id,
+            },
+        })
 
-    await prisma.watchedProduct.createMany({
-        data: body.watchedProductIds.map((e: number)=>{return  {organizationId: organization_?.id, productId: e}})
-    })
+        await prisma.watchedProduct.createMany({
+            data: body.watchedProductIds.map((e: number)=>{return  {organizationId: organization_?.id, productId: e}})
+        })
+    }
 
 
-    return new Response("Bad request", {status:400});
+    return {
+        status: "success",
+    };
 });
